+++ Dev/Editor/EffekseerCore/Data/RendererValues.cs
@@ -911,14 +911,14 @@ namespace Effekseer.Data
 
 				[IO(Export = true)]
 				[Key(key = "FalloffParameter_Pow")]
-				public Value.Int Pow { get; private set; }
+				public Value.Float Pow { get; private set; }
 
 				public FalloffParameter()
 				{
 					ColorBlendType = new Value.Enum<BlendType>(BlendType.Add);
 					BeginColor = new Value.Color(0, 0, 0, 255);
 					EndColor = new Value.Color(255, 255, 255, 255);
-					Pow = new Value.Int(1, 100, 1);
+					Pow = new Value.Float(1, 100, 1);
 				}
 
 				public enum BlendType : int
